requirements.txt
aiohttp
jinja2
s3fs
imageio[pyav]
requests
fsspec
intake
intake-xarray